prompt for filters

+++ ui/components/Filters.js
@@ -17,6 +17,7 @@ import UploadIcon from "@material-ui/icons/Publish";
 import { connect } from "react-redux";
 import { bindActionCreators } from "redux";
 import MUIDataTable from "mui-datatables";
+import PromptComponent from "./PromptComponent";
 import Moment from "react-moment";
 import { withSnackbar } from "notistack";
 import CloseIcon from "@material-ui/icons/Close";
@@ -106,6 +107,7 @@ function MesheryFilters({
   const [search] = useState("");
   const [sortOrder] = useState("");
   const [count, setCount] = useState(0);
+  const modalRef = useRef(null);
   const [pageSize, setPageSize] = useState(10);
   const [filters, setFilters] = useState([]);
   const [selectedRowData, setSelectedRowData] = useState(null);
@@ -371,7 +373,16 @@ function MesheryFilters({
     }
   });
 
-  async function deleteFilter(id) {
+  async function showmodal() {
+    let response = await modalRef.current.show({ title : "Delete Filter?",
+
+      subtitle : "Are you sure you want to delete this filter?",
+
+      options : ["YES", "NO"], })
+    return response;
+  }
+
+  function deleteFilter(id) {
     dataFetch(
       `/api/filter/${id}`,
       {
@@ -416,9 +427,15 @@ function MesheryFilters({
     download : false,
     customToolbar : CustomToolbar(uploadHandler, urlUploadHandler),
 
-    onRowsDelete : function handleDelete(row) {
-      const fid = Object.keys(row.lookup).map((idx) => filters[idx]?.id);
-      fid.forEach((fid) => deleteFilter(fid));
+    onRowsDelete : async function handleDelete(row) {
+      let response  = await showmodal()
+      console.log(response)
+      if (response === "YES") {
+        const fid = Object.keys(row.lookup).map((idx) => filters[idx]?.id);
+        fid.forEach((fid) => deleteFilter(fid));
+      }
+      if (response === "NO")
+        fetchFilters(page, pageSize, search, sortOrder);
     },
 
     onTableChange : (action, tableState) => {
@@ -473,6 +490,7 @@ function MesheryFilters({
         // @ts-ignore
         options={options}
       />
+      <PromptComponent ref={modalRef} />
     </NoSsr>
   );
 }
